test(sdpa): Rubin exact-LSE tests use an fp64 reference; assert_close_fp8_grad proves a single fp8 flip by its shape (sm107 CI after #1059) (#1075)
* test(sdpa): the Rubin exact-LSE tests build their reference in fp64; assert_close_fp8_grad proves a single fp8 flip by its shape

PR #1059's own sm107 pipeline (67882897, job 439855192) left four cases red on the
212-SM lane, none of them a kernel defect:

* `test_fp8_stats_is_the_exact_softmax_lse[*causal*]` x3 -- max |dLSE| 1.3e-4 against the
  1e-4 bound, rms 1e-5, IDENTICAL across the f32 and f16x2 exponent arms.  The arm-
  independence is the tell: the test's own `q.float() @ k.float().T` reference runs in TF32
  in the DLFW container (`TORCH_ALLOW_TF32_CUBLAS_OVERRIDE=1`; torch also defaults
  `allow_tf32=True` in this venv), ~3e-4 relative per logit.  A 1024-column log-sum-exp
  averages that to ~2e-5; a causal row with ONE valid column keeps it whole (measured on
  A100: TF32 1.9e-4 on row 0, true fp32 1.5e-7).  Both exact-LSE tests (fp8 and mxfp8)
  now build the logits from fp64 operands, as `fp8_ref.compute_ref_forward(dtype=float64)`
  already does for the harness.  The kernels are unchanged and were exact.

* `test_mhas_v2::test_sdpa_fp8_bwd_L0[test310]` (d192x128, SWA, FROST fwd + native bwd) --
  dK row 1464 moved UNIFORMLY by +-0.5 (147 of 192 elements outside tolerance), dQ row 575
  by <= 0.25: one flipped fp8 dS code at (575, 1464) seen through Q and through K.  Row 575
  is a negative-score q row, amplitude m = 8 at d192 / attn_scale 1/8, so one dS step of
  2^-4 x dP_descale moves the whole dK row by 0.5 -- above the harness's fixed `4 * atol`
  row cap (0.32), which was calibrated on amplitude-4 rows (the GB300 test310 flip, 0.25).
  Absent from the previous nightly and passing on the 208-SM dev node (Philox layout
  follows the SM count): #1059's exact Stats re-rolled every midpoint.

  Rather than raise the cap (which is what rejects a real defect -- #981 measured 0.42-2.0),
  `assert_close_fp8_grad` gains an opt-in structure check: with `operand=` (Q for dK, K for
  dQ, dO for dV, V for O) and `flip_unit=` (dP_descale / s_descale) every over-cap bad row
  must be ONE flipped intermediate -- a rank-1 update along a SINGLE operand row whose
  fitted step is a power of two times the descale, after which the row meets the ordinary
  tolerance.  A tile, a band, a garbage row, a non-power-of-two step or a row fed by two
  flips all fail it (verified on the dumped test310 dev-node tensors and pinned by the new
  `sdpa/test_fp8_flip_budget.py`, L0).  `exec_sdpa_fp8` passes the operands at its four
  compare sites; the element and row budgets are untouched.

test/AGENTS.md records both traps (fp64 references below ~1e-3; check a flip's shape
before touching a cap).  Validated on the sm107 dev node (208 SMs, cc 10.7): test310,
fp8_bwd_ragged test31, fp8_fwd test69 / test136, both exact-LSE tests and the self-test --
19 passed, 2 skipped; self-test 5/5 on A100 too.


* test(sdpa): the fp8 single-flip fallback needs intermediate-level evidence, not a row shape

Review of #1075 (Codex P1): "a power-of-two multiple of an operand row does not
establish that two computations rounded one intermediate to adjacent FP8 codes"
-- the shape-only fallback accepted a masked-key contribution, a cross-batch and
a cross-head row, two identical rows each flipped once, and a nominal e4m3 step
of 8192 (the widest adjacent e4m3 spacing is 32).

assert_close_fp8_grad(operand=, flip_unit=, intermediates=, kind=, fp8_dtype=,
out_dtype=) now keeps the 4*atol cap unless EVERY bad d-row is explained by ONE
flipped intermediate with evidence at the intermediate itself, gathered lazily
(only when the cap fails, only for the bad rows): the candidate lies in the row's
own reduction (same batch, the GQA group of its kv head, an unmasked position
per the reference's `valid`); the reference's scaled p / dS sits within 1/32 of a
spacing of a rounding midpoint between two adjacent codes of the intermediate's
fp8 dtype; the recomputed row (c_alt - c_ref) * descale * gain * operand_row
reproduces the deviation within the ordinary tolerance AND within half an
output code from each side; and the least-squares flip count is 1 +- 0.5 (a
1.5u / 2.5u deviation on a high-|expected| row otherwise hides inside
atol + rtol*|expected|).  fp8_ref.compute_ref / compute_ref_backward gain
return_intermediates= (default path a pure refactor: p_scaled / ds_scaled /
valid / gain for the requested rows only -- a full [b,h,s,s] matrix is 17 GB at
s=8192).  Ragged/packed layouts keep the plain cap.  The attribution (b, hq, i,
j), x, c_ref -> c_alt and the step are printed for every accepted row; a
rejected row prints the closest scalar-multiple fit and which piece of evidence
it lacks.

sdpa/test_fp8_flip_budget.py (44 cases, e4m3 + e5m2, reference-built problem):
the five review controls and the multi-flip bypass are REJECTED by the intended
condition; a genuine midpoint flip (incl. the test310 shape: dK row uniformly
+-0.5 = one dS flip x |q| = 8) is ACCEPTED with its attribution; the plain
budget + cap is unchanged without intermediates.  coderabbit: compute_ref_forward
-> compute_ref in test/AGENTS.md and test_sdpa_fp8_sm107.py.

GPU: SM100 4u4g-gen-0150 test_mhas_v2 -k fp8 + the self-tests: 821 passed, 1
failed = test_sdpa_mxfp8_fwd_L0[test241] "stats mismatch: 374 elements", which
fails identically on develop with FROST engines on and off (pre-existing, not
this change).  Rubin 212-SM vr-nvl72-ts4b-l11-130-c09: test_sdpa_fp8_bwd_L0[test310],
test_sdpa_fp8_bwd_ragged_L0[test31] and the 44 self-tests: 46 passed.


---------

+++ test/python/sdpa/frost/test_sdpa_fwd_dsl_sm107.py
@@ -985,7 +985,9 @@ def test_mxfp8_stats_is_the_exact_softmax_lse(d_qk, d_v, causal):
         data_d, sf_d, swz_d, data_s, sf_s, swz_s = quantize_to_mxfp8(x.contiguous(), b, h, s, d, 32, torch.float8_e4m3fn, with_ref=True)
         data, sf, swz = (data_s, sf_s, swz_s) if columnwise else (data_d, sf_d, swz_d)
         # sf_*_ref are the per-element fp32 DEQUANT SCALES [b, h, s, d]; the value the kernel sees is data * scale.
-        deq = data.float().reshape(b, h, s, d) * sf.float().reshape(b, h, s, d)
+        # fp64 so the reference logits below are exact: the DLFW containers run fp32 matmul in TF32, which a
+        # causal row with one valid column turns into a 1e-4-class LSE error (test_fp8_stats_is_the_exact_softmax_lse).
+        deq = data.double().reshape(b, h, s, d) * sf.double().reshape(b, h, s, d)
         return data.permute(0, 2, 1, 3).contiguous().transpose(1, 2), deq, swz.contiguous()
 
     q8, q_deq, sfq = mx(qf, hq, d_qk, False)
@@ -1008,7 +1010,7 @@ def test_mxfp8_stats_is_the_exact_softmax_lse(d_qk, d_v, causal):
     logits = q_deq @ k_deq.repeat_interleave(rep, 1).transpose(-1, -2) * d_qk**-0.5
     if causal:
         logits = logits.masked_fill(~torch.tril(torch.ones(s, s, dtype=torch.bool, device=dev)), float("-inf"))
-    lse_ref = torch.logsumexp(logits.double(), dim=-1).float()
+    lse_ref = torch.logsumexp(logits, dim=-1).float()
     assert torch.isfinite(lse).all(), "unwritten LSE rows"
     err = (lse - lse_ref).abs()
     assert (
